requirements_analysis.txt
access==1.1.9
adjustText==1.3.0
affine==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
anndata==0.10.4
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-api-compat==1.4
arrow==1.3.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autograd==1.6.2
autograd-gamma==0.5.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blosc2==2.5.1
cachetools==5.3.3
cattrs==23.2.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1704278392174/work
commonmark==0.9.1
contourpy==1.2.0
cuda-python==12.4.0
cudf-cu12==24.2.2
cugraph-cu12==24.2.0
cuml-cu12==24.2.0
cupy-cuda12x==13.0.0
cycler==0.12.1
Cython==3.0.7
dask==2024.1.1
dask-cuda==24.2.0
dask-cudf-cu12==24.2.2
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
decoupler==1.6.0
defusedxml==0.7.1
deprecation==2.1.0
distributed==2024.1.1
docstring-to-markdown==0.15
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
esda==2.6.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1700579780973/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fastcluster==1.2.6
fastjsonschema==2.19.1
fastrlock==0.8.2
fiona==1.10.1
fonttools==4.47.0
forestplot==0.4.0
formulaic==1.0.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
future==0.18.3
geopandas==1.0.1
giddy==2.3.5
great-tables==0.14.0
h11==0.14.0
h5py==3.10.0
hdmedians==0.14.2
htmltools==0.5.1
httpcore==1.0.5
httpx==0.27.0
idna==3.6
igraph==0.11.3
imagecodecs==2024.12.30
imageio==2.33.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
inequality==1.0.1
interface-meta==1.3.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1703631723894/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1703263236586/work
ipywidgets==8.1.2
isoduration==20.11.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
jedi-language-server==0.41.4
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1704321308630/work
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lazy_loader==0.3
legacy-api-wrap==1.4
leidenalg==0.10.2
libpysal==4.12.1
lifelines==0.28.0
llvmlite==0.41.1
locket==1.0.0
lsprotocol==2023.0.1
lxml==5.1.0
mapclassify==2.8.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mgwr==2.2.1
mistune==3.0.2
mizani==0.11.2
momepy==0.8.1
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
ndindex==1.8
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1697083700168/work
networkx==3.2.1
notebook==7.1.3
notebook_shim==0.2.4
numba==0.58.1
numdifftools==0.9.41
numexpr==2.9.0
numpy==1.24.4
nvtx==0.2.10
opencv-python==4.9.0.80
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==1.5.3
pandas-flavor==0.6.0
pandocfilters==1.5.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
partd==1.4.1
patsy==0.5.6
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.2.0
pingouin==0.5.4
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1701708255999/work
plotnine==0.13.5
pointpats==2.5.1
prometheus_client==0.20.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==4.25.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
PuLP==2.9.0
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py-cpuinfo==9.0.0
pyarrow==14.0.2
pybetareg==1.0.0
PyComplexHeatmap==1.7.3
pycparser==2.22
pygls==1.3.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pylibcugraph-cu12==24.2.0
pylibraft-cu12==24.2.0
pynndescent==0.5.11
pynvjitlink-cu12==0.1.14
pynvml==11.4.1
pyogrio==0.10.0
pyparsing==3.1.1
pyproj==3.7.0
pysal==24.7
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1686601365461/work
qtconsole==5.5.2
QtPy==2.4.1
quantecon==0.7.2
quarto-cli==1.5.57
raft-dask-cu12==24.2.0
rapids-dask-dependency==24.2.0
rapids_singlecell==0.10.1
rasterio==1.4.1
rasterstats==0.20.0
referencing==0.35.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rmm-cu12==24.2.0
rpds-py==0.18.1
Rtree==1.3.0
scanpy==1.10.0
scikit-bio==0.5.9
scikit-fuzzy==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.0
scikit-misc==0.3.1
scikit-posthocs==0.9.0
scipy==1.13.0
seaborn==0.13.1
segregation==2.5
Send2Trash==1.8.3
session-info==1.0.0
shap==0.46.0
shapely==2.0.6
simplejson==3.19.3
sinfo==0.3.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
slicer==0.0.8
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spaghetti==1.7.6
spglm==1.1.0
spint==1.0.7
splot==1.1.7
spopt==0.6.1
spreg==1.7
spvcm==0.3.0
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
statsmodels==0.14.1
stdlib-list==0.10.0
style==1.1.0
sympy==1.13.3
tables==3.9.2
tabulate==0.9.0
tblib==3.0.0
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.2.0
tifffile==2023.12.9
tinycss2==1.3.0
tobler==0.11.3
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1704212992681/work
treelite==4.0.0
types-python-dateutil==2.9.0.20240316
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1702176139754/work
tzdata==2023.4
ucx-py-cu12==0.36.0
umap-learn==0.5.5
update==0.0.1
uri-template==1.3.0
urllib3==2.1.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1700607916581/work
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.10
wrapt==1.16.0
xarray==2024.1.1
yarl==1.9.4
zict==3.0.0
zipp==3.18.1
